Windows port: op-family MSVC fixes (narrowing casts, builtins, force-inline)
- explicit casts where MSVC hard-errors on narrowing in braced initializers
  (GNU/Clang builds run these files with -Wno-narrowing): data_movement
  (concat, copy, gather, indexed_fill, moe remaps, roll, scatter, sort,
  transpose), eltwise (binary_ng, ternary, unary), matmul factories,
  reduction (accumulation, topk), core work_split
- __builtin_unreachable() call sites routed through ttsl::unreachable()
- __builtin_clz replaced with std::bit_floor in topk_utils
- bare glibc 'uint' spelled uint32_t in unary_op_utils
- slice_device_operation uses TT_FORCE_INLINE instead of raw
  __attribute__((always_inline))

+++ ttnn/cpp/ttnn/operations/data_movement/common/common.cpp
@@ -471,8 +471,8 @@ ttnn::Tensor pad_to_tile_vol(
         uint32_t num_non_hw_dims = rank - 2u;
         auto padding_vec = ttsl::SmallVector<std::array<uint32_t, 2>>(num_non_hw_dims, {0, 0});
         padding_vec.reserve(rank);
-        padding_vec.emplace_back(0, padded_height - padded_shape[-2]);
-        padding_vec.emplace_back(0, padded_width - padded_shape[-1]);
+        padding_vec.emplace_back(0u, static_cast<uint32_t>(padded_height - padded_shape[-2]));
+        padding_vec.emplace_back(0u, static_cast<uint32_t>(padded_width - padded_shape[-1]));
 
         auto padded_output = ttnn::pad(tensor, padding_vec, value, use_multicore, memory_config);
         TT_FATAL(
